sample adapter method for fetching project ids for list of samples

+++ test/dbadaptor/sampleadaptor_test.py
@@ -13,55 +13,60 @@ from igf_data.igfdb.runadaptor import RunAdaptor
 
 class Sampleadaptor_test1(unittest.TestCase):
   def setUp(self):
-    self.dbconfig='data/dbconfig.json'
-    dbparam=read_dbconf_json(self.dbconfig)
-    base=BaseAdaptor(**dbparam)
-    self.engine=base.engine
-    self.dbname=dbparam['dbname']
+    self.dbconfig = 'data/dbconfig.json'
+    dbparam = read_dbconf_json(self.dbconfig)
+    base = BaseAdaptor(**dbparam)
+    self.engine = base.engine
+    self.dbname = dbparam['dbname']
     Base.metadata.create_all(self.engine)
-    self.session_class=base.get_session_class()
-    project_data=[{'project_igf_id':'IGFP0001_test_22-8-2017_rna',
-                   'project_name':'test_22-8-2017_rna',
-                   'description':'Its project 1',
-                   'project_deadline':'Before August 2017',
-                   'comments':'Some samples are treated with drug X',
-                 }]
+    self.session_class = base.get_session_class()
+    project_data = [{
+      'project_igf_id':'IGFP0001_test_22-8-2017_rna',
+      'project_name':'test_22-8-2017_rna',
+      'description':'Its project 1',
+      'project_deadline':'Before August 2017',
+      'comments':'Some samples are treated with drug X'}]
     base.start_session()
-    pa=ProjectAdaptor(**{'session':base.session})
+    pa = ProjectAdaptor(**{'session':base.session})
     pa.store_project_and_attribute_data(data=project_data)
     base.close_session()
 
   def tearDown(self):
     Base.metadata.drop_all(self.engine)
-    os.remove(self.dbname)
+    if os.path.exists(self.dbname):
+      os.remove(self.dbname)
 
   def test_store_sample_and_attribute_data(self):
-    sa=SampleAdaptor(**{'session_class': self.session_class})
-    sample_data=[{'sample_igf_id':'IGFS001','library_id':'IGFS001','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS002','library_id':'IGFS002','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS003','library_id':'IGFS003','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS004','library_id':'IGFS004','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                ]
+    sa = SampleAdaptor(**{'session_class': self.session_class})
+    sample_data = [{
+      'sample_igf_id':'IGFS001','library_id':'IGFS001','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS002','library_id':'IGFS002','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS003','library_id':'IGFS003','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS004','library_id':'IGFS004','project_igf_id':'IGFP0001_test_22-8-2017_rna'}]
     sa.start_session()
     sa.store_sample_and_attribute_data(data=sample_data)
-    sa1=sa.check_sample_records_igf_id(sample_igf_id='IGFS001')
+    sa1 = sa.check_sample_records_igf_id(sample_igf_id='IGFS001')
     sa.close_session()
     self.assertEqual(sa1,True)
 
   def test_check_project_and_sample(self):
-    sa=SampleAdaptor(**{'session_class': self.session_class})
-    sample_data=[{'sample_igf_id':'IGFS001','library_id':'IGFS001','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS002','library_id':'IGFS002','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS003','library_id':'IGFS003','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                 {'sample_igf_id':'IGFS004','library_id':'IGFS004','project_igf_id':'IGFP0001_test_22-8-2017_rna',},
-                ]
+    sa = SampleAdaptor(**{'session_class': self.session_class})
+    sample_data = [{
+      'sample_igf_id':'IGFS001','library_id':'IGFS001','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS002','library_id':'IGFS002','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS003','library_id':'IGFS003','project_igf_id':'IGFP0001_test_22-8-2017_rna'},{
+      'sample_igf_id':'IGFS004','library_id':'IGFS004','project_igf_id':'IGFP0001_test_22-8-2017_rna'}]
     sa.start_session()
     sa.store_sample_and_attribute_data(data=sample_data)
-    sa1=sa.check_project_and_sample(project_igf_id='IGFP0001_test_22-8-2017_rna',\
-                                    sample_igf_id='IGFS001')
+    sa1 = \
+      sa.check_project_and_sample(
+        project_igf_id='IGFP0001_test_22-8-2017_rna',
+        sample_igf_id='IGFS001')
     self.assertEqual(sa1,True)
-    sa2=sa.check_project_and_sample(project_igf_id='IGFP0001_test_22-8-2017_rna',\
-                                    sample_igf_id='IGFS0011')
+    sa2 = \
+      sa.check_project_and_sample(
+        project_igf_id='IGFP0001_test_22-8-2017_rna',
+        sample_igf_id='IGFS0011')
     self.assertEqual(sa2,False)
     sa.close_session()
 
@@ -128,7 +133,8 @@ class Sampleadaptor_test2(unittest.TestCase):
 
   def tearDown(self):
     Base.metadata.drop_all(self.engine)
-    os.remove(self.dbname)
+    if os.path.exists(self.dbname):
+      os.remove(self.dbname)
 
   def test_fetch_seqrun_and_platform_list_for_sample_id(self):
     sa = SampleAdaptor(**{'session_class': self.session_class})
@@ -174,7 +180,8 @@ class Sampleadaptor_test3(unittest.TestCase):
 
   def tearDown(self):
     Base.metadata.drop_all(self.engine)
-    os.remove(self.dbname)
+    if os.path.exists(self.dbname):
+      os.remove(self.dbname)
 
   def test_fetch_sample_species_name(self):
     sa = SampleAdaptor(**{'session_class': self.session_class})
@@ -190,6 +197,63 @@ class Sampleadaptor_test3(unittest.TestCase):
     self.assertTrue(species_name is None)
     sa.close_session()
 
+class Sampleadaptor_test4(unittest.TestCase):
+  def setUp(self):
+    self.dbconfig = 'data/dbconfig.json'
+    dbparam = read_dbconf_json(self.dbconfig)
+    base = BaseAdaptor(**dbparam)
+    self.engine = base.engine
+    self.dbname = dbparam['dbname']
+    Base.metadata.create_all(self.engine)
+    self.session_class = base.get_session_class()
+    base.start_session()
+    project_data = [{
+      'project_igf_id':'IGFP0001_test_22-8-2017_rna',
+      'project_name':'test_22-8-2017_rna',
+      'description':'Its project 1',
+      'project_deadline':'Before August 2017',
+      'comments':'Some samples are treated with drug X'
+      },{
+      'project_igf_id':'IGFP0002_test_22-8-2017_rna',
+      'project_name':'test_22-8-2017_rna2'}]
+    pa = ProjectAdaptor(**{'session':base.session})
+    pa.store_project_and_attribute_data(data=project_data)
+    sample_data = [{
+      'sample_igf_id':'IGF00001',
+      'project_igf_id':'IGFP0001_test_22-8-2017_rna',
+      'species_name':'HG38'},{
+      'sample_igf_id':'IGF00002',
+      'project_igf_id':'IGFP0001_test_22-8-2017_rna',
+      'species_name':'UNKNOWN'},{
+      'sample_igf_id':'IGF00003',
+      'project_igf_id':'IGFP0002_test_22-8-2017_rna'}]
+    sa = SampleAdaptor(**{'session':base.session})
+    sa.store_sample_and_attribute_data(data=sample_data)
+    base.close_session()
+
+  def tearDown(self):
+    Base.metadata.drop_all(self.engine)
+    if os.path.exists(self.dbname):
+      os.remove(self.dbname)
+
+  def test_get_project_ids_for_list_of_samples(self):
+    sa = SampleAdaptor(**{'session_class': self.session_class})
+    sa.start_session()
+    project_ids = \
+      sa.get_project_ids_for_list_of_samples(
+       ['IGF00001'])
+    self.assertTrue('IGFP0001_test_22-8-2017_rna' in project_ids)
+    project_ids = \
+      sa.get_project_ids_for_list_of_samples(
+       ['IGF00001','IGF00002'])
+    self.assertTrue('IGFP0001_test_22-8-2017_rna' in project_ids)
+    self.assertEqual(len(project_ids),1)
+    project_ids = \
+      sa.get_project_ids_for_list_of_samples(
+       ['IGF00001','IGF00003'])
+    self.assertTrue('IGFP0001_test_22-8-2017_rna' in project_ids)
+    self.assertTrue('IGFP0002_test_22-8-2017_rna' in project_ids)
+    self.assertEqual(len(project_ids),2)
